requirements.txt
Django==1.8.3
Fabric==1.10.0
argparse==1.2.1
django-allauth==0.22.0
django-modeltranslation==0.10
django-tinymce==2.0.1
django-widget-tweaks==1.4.1
ecdsa==0.11
oauthlib==1.0.0
paramiko==1.15.1
psycopg2==2.5.4
pycrypto==2.6.1
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
wsgiref==0.1.2
